Extract startup OAuth callbacks
