+++ CLAUDE.md
@@ -22,7 +22,7 @@ Web app (PWA) only: production runtime ships with no build system, no bundler, a
 - **`index.html`** — HTML skeleton (header, sidebar, modals, chat panel, script/CSS includes)
 - **`styles.css`** — all CSS (dark/light themes, 10 responsive breakpoints, touch/hover media queries)
 - **`js/`** — ES modules loaded via `js/main.js` (count drifts; `ls js/*.js | wc -l`). Grouped by concern below; read the source for exported symbols.
-  - **Core**: `state.js` (single mutable `state`), `utils.js` (escape, status, notifications, dialogs), `theme.js`, `constants.js` (option arrays, personalities, EMF presets), `schema.js` (`MARKER_SCHEMA`, unit conversions, pricing, EMF thresholds, usage tracking), `main.js` (entry point, OAuth callback, init), `startup-profile.js` (profile migration/load/display bootstrap), `app-event-listeners.js` (global modal/keyboard/refresh wiring)
+  - **Core**: `state.js` (single mutable `state`), `utils.js` (escape, status, notifications, dialogs), `theme.js`, `constants.js` (option arrays, personalities, EMF presets), `schema.js` (`MARKER_SCHEMA`, unit conversions, pricing, EMF thresholds, usage tracking), `main.js` (entry point/init), `startup-profile.js` (profile migration/load/display bootstrap), `startup-oauth-callbacks.js` (wearable/OpenRouter callback routing), `app-event-listeners.js` (global modal/keyboard/refresh wiring)
   - **Data pipeline**: `data.js` (`getActiveData` central pipeline), `profile.js` (CRUD + migrations), `adapters.js` (217 specialty markers across fattyAcids/metabolomix/oat/biostarks), `pdf-import.js`, `import-file-input.js` (lazy file-picker import routing), `import-drop-zone.js` (lazy import drop-zone binding), `export.js`, `backup.js`, `crypto.js` (AES-256-GCM + BroadcastChannel cross-tab), `pii.js` (regex + local-AI obfuscation with streaming sanitizer)
   - **AI + providers**: `api.js` (6 providers, `callClaudeAPI` router, OpenRouter OAuth PKCE, Venice E2EE branch, recommended-model tiering), `provider-panels.js` (extracted from settings.js), `cashu-wallet.js` (Cashu BIP-39 seed, Lightning fund/withdraw, fee splitting), `nostr-discovery.js` (Routstr node discovery via Kind 38421)
   - **Views + UI**: `views-router.js` (route validation, per-profile last-view restore, scroll anchoring), `dashboard-view-composition.js` (dashboard route/widget/control wiring), `dashboard-page-view.js` (dashboard route shell, empty-state onboarding, mobile handoff, widget page composition), `lens-pages.js` (Labs/Genome/Body/Insight/Recommendations page renderers), `lens-page-shell.js` (shared lens header, widget chrome, ordering, and dashboard toggles), `dashboard-widgets.js` (dashboard widget registry, defaults, per-profile widget prefs), `dashboard-widget-controls.js` (widget picker, layout actions, drag/reorder controls), `dashboard-widget-renderers.js` (dashboard widget body renderers and recommendation helpers), `recommendation-actions.js` (recommendation modal/chat/save/dismiss handlers), `category-page-view.js` (category route shell, view switching, card-order preservation), `category-view-renderers.js` (category chart cards, table/heatmap shells, fatty-acid profile renderers), `category-customization.js` (category/marker rename helpers and category icon picker), `light-page-view.js` (Light & Sun page shell, Light Today strip, dashboard Light pills), `light-channel-view.js` (Light channel pills, sparklines, drill-down panels, citations, and suggestions), `views.js` (route wiring, compatibility exports, compare, correlations, detail modal, manual entry, focus card, onboarding), `nav.js` (sidebar, avatar colors), `charts.js` (Chart.js plugins), `context-cards.js` (9 cards with AI health dots + interpretive lens + change history), `client-list.js`, `supplements.js` (+ `supplement-warnings.js`, `food-contaminants.js`), `recommendations.js` (DNA-aware catalog, 3 touchpoints), `cycle.js`, `emf.js` (Baubiologie SBM-2015), `feedback.js`, `tour.js`, `changelog.js`, `notes.js`, `hardware.js` (GPU detection, model fitness), `image-utils.js`
