+++ tests/test-correctness-phase2.js
@@ -77,6 +77,7 @@ const pwaAppShellAssets = [
   '/vendor/evolu/sqlite3.wasm',
   '/js/app-event-listeners.js',
   '/js/startup-profile.js',
+  '/js/startup-oauth-callbacks.js',
   '/js/ai-verdict-engine.js',
   '/js/import-loader.js',
   '/js/import-file-input.js',
