+++ tests/verify-modules.js
@@ -562,6 +562,9 @@
     const hasStartupProfileJs = sw.includes('/js/startup-profile.js');
     assert('Service worker caches js/startup-profile.js', hasStartupProfileJs);
 
+    const hasStartupOAuthCallbacksJs = sw.includes('/js/startup-oauth-callbacks.js');
+    assert('Service worker caches js/startup-oauth-callbacks.js', hasStartupOAuthCallbacksJs);
+
     const hasViewsJs = sw.includes('/js/views.js');
     assert('Service worker caches js/views.js', hasViewsJs);
